Move RawJtagIo
